AK+LibC+regex: Move C++ classes to AK and let LibC use the AK API
Improved little things:
- Use enum classes instead of enums
- Add string view of the match to match result
- Use isxxx functions/macros from ctype instead of error prone if conditions
