+++ Libraries/LibC/regex.h
@@ -26,18 +26,15 @@
 
 #pragma once
 
-#include <AK/Forward.h>
-#include <AK/HashMap.h>
 #include <AK/OwnPtr.h>
-#include <AK/String.h>
-#include <AK/Vector.h>
+#include <AK/Regex.h>
 #include <stddef.h>
 
-//#define REGEX_DEBUG
-
+namespace AK {
 namespace regex {
 class VM;
 class Token;
+}
 }
 
 __BEGIN_DECLS
@@ -66,7 +63,7 @@ struct regex_t {
     u8 cflags;
     u8 eflags;
     size_t re_minlength;
-    OwnPtr<regex::VM> vm;
+    OwnPtr<AK::regex::VM> vm;
     size_t re_pat_errpos { 0 };
     ReError re_pat_err;
     String re_pat;
@@ -81,19 +78,17 @@ struct regmatch_t {
 };
 
 // Values for the cflags parameter to the regcomp() function:
-#define REG_EXTENDED 1                  // Use Extended Regular Expressions.
-#define REG_ICASE (REG_EXTENDED << 1)   // Ignore case in match.
-#define REG_NOSUB (REG_EXTENDED << 2)   // Report only success or fail in regexec().
-#define REG_NEWLINE (REG_EXTENDED << 3) // Change the handling of newline.
+#define REG_EXTENDED (u8) AK::regex::CompilationFlags::Extended      // Use Extended Regular Expressions.
+#define REG_ICASE (u8) AK::regex::CompilationFlags::IgnoreCase       // Ignore case in match.
+#define REG_NOSUB (u8) AK::regex::CompilationFlags::NoSubExpressions // Report only success or fail in regexec().
+#define REG_NEWLINE (u8) AK::regex::CompilationFlags::HandleNewLine  // Change the handling of newline.
 
 // Values for the eflags parameter to the regexec() function:
-#define REG_NOTBOL 1                   // The circumflex character (^), when taken as a special character, will not match the beginning of string.
-#define REG_NOTEOL (REG_NOTBOL << 1)   // The dollar sign ($), when taken as a special character, will not match the end of string.
-#define REG_MATCHALL (REG_NOTBOL << 2) // Match all occurences of the character - extension to posix
-#define REG_SEARCH (REG_NOTBOL << 3)   // Do try to match the pattern anyhwere in the string
-#define REG_STATS (REG_NOTBOL << 4)    // Print stats for a match to stdout
-
-#define REG_MAX_RECURSE 5000
+#define REG_NOTBOL (u8) AK::regex::MatchFlags::NoBeginOfLine // The circumflex character (^), when taken as a special character, will not match the beginning of string.
+#define REG_NOTEOL (u8) AK::regex::MatchFlags::NoEndOfLine   // The dollar sign ($), when taken as a special character, will not match the end of string.
+#define REG_MATCHALL (u8) AK::regex::MatchFlags::MatchAll    // Match all occurences of the character - extension to posix
+#define REG_SEARCH (u8) AK::regex::MatchFlags::Search        // Do try to match the pattern anyhwere in the string
+#define REG_STATS (u8) AK::regex::MatchFlags::Stats          // Print stats for a match to stdout
 
 int regcomp(regex_t*, const char*, int);
 int regexec(const regex_t*, const char*, size_t, regmatch_t[], int);
@@ -101,282 +96,3 @@ size_t regerror(int, const regex_t*, char*, size_t);
 void regfree(regex_t*);
 
 __END_DECLS
-
-namespace regex {
-
-#define ENUMERATE_OPCODES              \
-    __ENUMERATE_OPCODE(Compare)        \
-    __ENUMERATE_OPCODE(Jump)           \
-    __ENUMERATE_OPCODE(ForkJump)       \
-    __ENUMERATE_OPCODE(ForkStay)       \
-    __ENUMERATE_OPCODE(SaveLeftGroup)  \
-    __ENUMERATE_OPCODE(SaveRightGroup) \
-    __ENUMERATE_OPCODE(CheckBegin)     \
-    __ENUMERATE_OPCODE(CheckEnd)       \
-    __ENUMERATE_OPCODE(Exit)
-
-enum class OpCode {
-#define __ENUMERATE_OPCODE(x) x,
-    ENUMERATE_OPCODES
-#undef __ENUMERATE_OPCODE
-};
-
-enum class CompareType {
-    Undefined,
-    Inverse,
-    AnySingleCharacter,
-    OrdinaryCharacter,
-    OrdinaryCharacters,
-    CharacterClass,
-    RangeExpression,
-    RangeExpressionDummy,
-};
-
-enum class CharacterClass {
-    Alnum,
-    Cntrl,
-    Lower,
-    Space,
-    Alpha,
-    Digit,
-    Print,
-    Upper,
-    Blank,
-    Graph,
-    Punct,
-    Xdigit,
-};
-
-class StackValue {
-public:
-    union CompareValue {
-        CompareValue(const CharacterClass value)
-            : character_class(value)
-        {
-        }
-        CompareValue(const char value1, const char value2)
-            : range_values { value1, value2 }
-        {
-        }
-        const CharacterClass character_class;
-        const struct {
-            const char from;
-            const char to;
-        } range_values;
-    };
-
-    union {
-        const OpCode op_code;
-        const char* string;
-        const char ch;
-        const int number;
-        const size_t positive_number;
-        const CompareValue compare_value;
-        const CompareType compare_type;
-    };
-
-    const char* name() const;
-    static const char* name(OpCode);
-
-    StackValue(const OpCode value)
-        : op_code(value)
-    {
-    }
-    StackValue(const char* value)
-        : string(value)
-    {
-    }
-    StackValue(const char value)
-        : ch(value)
-    {
-    }
-    StackValue(const int value)
-        : number(value)
-    {
-    }
-    StackValue(const size_t value)
-        : positive_number(value)
-    {
-    }
-    StackValue(const CharacterClass value)
-        : compare_value(value)
-    {
-    }
-    StackValue(const char value1, const char value2)
-        : compare_value(value1, value2)
-    {
-    }
-    StackValue(const CompareType value)
-        : compare_type(value)
-    {
-    }
-
-    ~StackValue() = default;
-};
-
-struct CompareTypeAndValue {
-    CompareType type;
-    StackValue value;
-};
-
-#define ENUMERATE_REGEX_TOKENS                 \
-    __ENUMERATE_REGEX_TOKEN(Eof)               \
-    __ENUMERATE_REGEX_TOKEN(OrdinaryCharacter) \
-    __ENUMERATE_REGEX_TOKEN(Circumflex)        \
-    __ENUMERATE_REGEX_TOKEN(Period)            \
-    __ENUMERATE_REGEX_TOKEN(LeftParen)         \
-    __ENUMERATE_REGEX_TOKEN(RightParen)        \
-    __ENUMERATE_REGEX_TOKEN(LeftCurly)         \
-    __ENUMERATE_REGEX_TOKEN(RightCurly)        \
-    __ENUMERATE_REGEX_TOKEN(LeftBracket)       \
-    __ENUMERATE_REGEX_TOKEN(RightBracket)      \
-    __ENUMERATE_REGEX_TOKEN(Asterisk)          \
-    __ENUMERATE_REGEX_TOKEN(EscapeSequence)    \
-    __ENUMERATE_REGEX_TOKEN(Dollar)            \
-    __ENUMERATE_REGEX_TOKEN(Pipe)              \
-    __ENUMERATE_REGEX_TOKEN(Plus)              \
-    __ENUMERATE_REGEX_TOKEN(Comma)             \
-    __ENUMERATE_REGEX_TOKEN(Questionmark)
-
-enum class TokenType {
-#define __ENUMERATE_REGEX_TOKEN(x) x,
-    ENUMERATE_REGEX_TOKENS
-#undef __ENUMERATE_REGEX_TOKEN
-};
-
-class Token {
-public:
-    Token(TokenType type, size_t start_position, StringView value)
-        : m_type(type)
-        , m_position(start_position)
-        , m_value(value)
-    {
-    }
-
-    TokenType type() const { return m_type; }
-    const char* name() const;
-    static const char* name(TokenType);
-    const StringView& value() const { return m_value; }
-    size_t position() const { return m_position; }
-
-private:
-    TokenType m_type;
-    size_t m_position;
-    StringView m_value;
-};
-
-class Lexer {
-public:
-    explicit Lexer(StringView source);
-    Token next();
-    void reset();
-    void back(size_t offset);
-
-private:
-    char peek(size_t offset = 0) const;
-    void consume();
-
-    StringView m_source;
-    size_t m_position = 0;
-    size_t m_previous_position = 0;
-    Token m_current_token;
-    int m_current_char;
-};
-
-class Parser {
-public:
-    explicit Parser(Lexer lexer);
-
-    struct ParserResult {
-        Vector<StackValue> m_bytes;
-        size_t m_match_groups;
-        size_t m_min_match_length;
-        ReError m_error;
-        Token m_error_token;
-    };
-
-    ParserResult parse(int cflags);
-    bool has_error() const { return m_parser_state.m_error != REG_NOERR; }
-
-private:
-    bool match(TokenType type) const;
-    bool match(char ch) const;
-    bool done() const;
-    void expected(const char* what);
-    Token consume();
-    Token consume(TokenType type);
-    bool consume(StringView view);
-
-    void save_state();
-    void load_state();
-    bool match_meta_chars();
-    bool match_ere_quoted_chars();
-    bool match_ere_dupl_symbol();
-    bool parse_ere_dupl_symbol(Vector<StackValue>&, size_t& min_length);
-    bool parse_ere_expression(Vector<StackValue>&, size_t& min_length);
-    bool parse_extended_reg_exp(Vector<StackValue>&, size_t& min_length);
-    bool parse_bracket_expression(Vector<StackValue>&, size_t& min_length);
-    void reset();
-
-    bool set_error(ReError error);
-
-    struct ParserState {
-        Lexer m_lexer;
-        Token m_current_token;
-        ReError m_error = REG_NOERR;
-        Token m_error_token { TokenType::Eof, 0, StringView(nullptr) };
-        Vector<StackValue> m_bytes;
-        size_t m_match_groups { 0 };
-        size_t m_min_match_length { 0 };
-        int m_cflags { 0 };
-        explicit ParserState(Lexer);
-    };
-
-    ParserState m_parser_state;
-    Optional<ParserState> m_saved_state;
-};
-
-class VM {
-public:
-    explicit VM(const Vector<StackValue>& bytecode, const String&& pattern, const int cflags)
-        : m_bytecode(bytecode)
-        , m_pattern(move(pattern))
-        , m_cflags(cflags) {};
-
-    struct MatchResult {
-        size_t m_match_count { 0 };
-        Vector<regmatch_t> m_matches {};
-        size_t m_ops { 0 };
-    };
-
-    MatchResult match(const StringView view, size_t max_matches_result, size_t match_groups, size_t min_length, int flags) const;
-    const Vector<StackValue>& bytes() const { return m_bytecode; }
-
-private:
-    struct ForkStayTuple {
-        size_t m_instructionp { 0 };
-        size_t m_stringp { 0 };
-    };
-
-    struct MatchState {
-        StringView m_view;
-        size_t m_instructionp { 0 };
-        size_t m_stringp { 0 };
-        size_t m_ops { 0 };
-        size_t m_matches_offset { 0 };
-        Vector<regmatch_t> m_matches;
-        Vector<regoff_t> m_left;
-        int m_eflags;
-
-        MatchState() = default;
-    };
-
-    bool match_recurse(MatchState& state, size_t recursion_level = 0) const;
-    const StackValue get(MatchState& state, size_t offset = 0) const;
-    const StackValue get_and_increment(MatchState& state, size_t value = 1) const;
-
-    const Vector<StackValue> m_bytecode;
-    const String& m_pattern;
-    const int m_cflags { 0 };
-};
-}
